Switch to webpack, which allows for easier builds to firefox Also allows use of SASS
