--- css/mini.css
@@ -1,38 +1,0 @@
-body[betteryt-mini] .html5-main-video {
-  top: 0px !important;
-  left: 0px !important;
-  width: var(--ytd-miniplayer-width) !important;
-  height: var(--ytd-miniplayer-height) !important;
-  object-fit: contain !important;
-}
-
-body[betteryt-mini] .ytp-chrome-controls {
-  width: calc(var(--ytd-miniplayer-width) - 2px);
-}
-
-body[betteryt-mini] .ytp-chrome-bottom {
-  left: 1px !important;
-}
-
-body[betteryt-mini]
-  .ytp-left-controls
-  > *:not(.ytp-play-button, .ytp-volume-area, .ytp-time-display),
-body[betteryt-mini] .ytp-right-controls > *:not(.ytp-miniplayer-button),
-body[betteryt-mini]
-  #movie_player
-  > *:not(.html5-video-container, .ytp-chrome-bottom, .betteryt),
-body[betteryt-mini] .ytp-chrome-bottom .ytp-progress-bar-container,
-body[betteryt-mini] .expander.ytd-miniplayer,
-body[betteryt-mini] .index-message.ytd-miniplayer {
-  display: none !important;
-}
-
-body[betteryt-mini] .ytp-miniplayer-button {
-  display: unset !important;
-}
-
-body[betteryt-live] .betteryt.ytp-progress-bar-container,
-body:not([betteryt-mini]) .betteryt.ytp-progress-bar-container,
-body:not([betteryt-mini]) .betteryt.ytp-gradient-bottom {
-  display: none;
-}
