--- css/pip.css
@@ -1,3 +1,0 @@
-body[betteryt-pip] .ytp-pip-button {
-  display: unset !important;
-}
