+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "betteryt",
+  "version": "0.0.7",
+  "description": "",
+  "private": true,
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "npm run start:chrome",
+    "start:chrome": "npx webpack --env browser=chrome --mode=development --devtool=inline-source-map --watch",
+    "start:firefox": "npx webpack --env browser=firefox --mode=development --devtool=inline-source-map --watch",
+    "build": "npm run build:chrome",
+    "build:chrome": "npx webpack --env browser=chrome",
+    "build:firefox": "npx webpack --env browser=firefox"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.7.1",
+    "mini-css-extract-plugin": "^2.6.1",
+    "sass": "^1.54.0",
+    "sass-loader": "^13.0.2",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0"
+  }
+}
