+++ src/_locales/fr/messages.json
