+++ src/css/mini_player.scss
@@ -1,0 +1,44 @@
+body[betteryt-mini] {
+  .html5-main-video {
+    top: 0px !important;
+    left: 0px !important;
+    width: var(--ytd-miniplayer-width) !important;
+    height: var(--ytd-miniplayer-height) !important;
+    object-fit: contain !important;
+  }
+
+  .ytp-chrome-controls {
+    width: calc(var(--ytd-miniplayer-width) - 2px);
+  }
+
+  .ytp-chrome-bottom {
+    left: 1px !important;
+  }
+
+  .ytp-left-controls
+    > *:not(.ytp-play-button, .ytp-volume-area, .ytp-time-display),
+  .ytp-right-controls > *:not(.ytp-miniplayer-button),
+  #movie_player > *:not(.html5-video-container, .ytp-chrome-bottom, .betteryt),
+  .ytp-chrome-bottom .ytp-progress-bar-container,
+  .expander.ytd-miniplayer,
+  .index-message.ytd-miniplayer {
+    display: none !important;
+  }
+
+  .ytp-miniplayer-button {
+    display: unset !important;
+  }
+}
+
+body:not([betteryt-mini]) {
+  .betteryt.ytp-progress-bar-container,
+  .betteryt.ytp-gradient-bottom {
+    display: none;
+  }
+}
+
+body[betteryt-live] {
+  .betteryt.ytp-progress-bar-container {
+    display: none;
+  }
+}
