+++ src/css/pip.scss
@@ -1,0 +1,5 @@
+body[betteryt-pip] {
+  .ytp-pip-button {
+    display: unset !important;
+  }
+}
