+++ src/html/popup.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <link rel="stylesheet" href="/css/popup.css" />
+    <link rel="stylesheet" href="../css/popup.css" />
   </head>
 
   <body>
@@ -92,7 +92,6 @@
 
     <p class="sub-label refresh" data-locale="settingsRefresh"></p>
 
-    <script src="/js/utils.js"></script>
-    <script src="/js/popup.js"></script>
+    <script src="../js/popup.js"></script>
   </body>
 </html>
