+++ src/images/toolbar-icon-32.png
