+++ src/js/modules/pip.js
@@ -1,0 +1,3 @@
+const testVideo = document.createElement('video');
+if (testVideo.requestPictureInPicture)
+  document.body.setAttribute('betteryt-pip', '');
